+++ dotNet5783_2774_6645/DalList/DalOrderItem.cs
@@ -14,7 +14,7 @@ internal class DalOrderItem : IOrderItem
     public int Add(OrderItem o)
     {
         DataSource.OrderItemList.Add(o);
-        return o.OrderID;
+        return (int)o.OrderID;
     }
 
     /// <summary>
